requirements.txt
flask
python-nmap
requests
supabase
python-dotenv
